requirements.txt
fastapi
uvicorn
pinecone-client>=3.0.0
sentence-transformers
python-dotenv
openai
pinecone
